# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Stoxtech Admin API",
        "version": "1.0.0"
    },
    "paths": {
...
        "/admin/brands/{id}/db/import": {
            "post": {
                "tags": [
                    "Brand"
                ],
                "summary": "Import db brand",
                "description": "Import tables into brand db.",
                "operationId": "brands.db.import",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "modules"
                                ],
                                "properties": {
                                    "modules": {
                                        "type": "array",
                                        "items": {
                                            "type": "string"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
        "/admin/auth/forget-password": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Send Reset Password Notification",
                "operationId": "6588ee88df2cbd2ec882d00b8ec24eb2",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "202": {
                        "description": "Accepted"
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationError"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                }
            }
        },
        "/admin/auth/reset-password": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Reset Password",
                "operationId": "f90e47572b7c88223f3c144e7ca73026",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password",
                                    "password_confirmation",
                                    "token"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password"
                                    },
                                    "password_confirmation": {
                                        "type": "string",
                                        "format": "password"
                                    },
                                    "token": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "202": {
                        "description": "Accepted"
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationError"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                }
            }
        },
        "/admin/auth/callback/{provider}": {
            "get": {
                "tags": [
                    "Auth"
                ],
                "summary": "OAuth provider callback endpoint",
                "operationId": "7e90017f737dac29a1d287d87bc7cc15",
                "parameters": [
                    {
                        "name": "provider",
                        "in": "path",
                        "description": "OAuth provider name",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "google"
                        }
                    }
                ],
                "responses": {
                    "302": {
                        "description": "Redirects to `redirect_url` with cookies (and possibly validation or application errors in query params)",
                        "headers": {
                            "Set-Cookie": {
                                "schema": {
                                    "type": "string",
                                    "example": "laravel_session=eyJpdiI6IjZKZm...%3D; Path=/; Domain=localhost; HttpOnly; Expires=Fri, 18 Nov 2022 13:25:26 GMT;"
                                }
                            },
                            "Location": {
                                "schema": {
                                    "type": "string",
                                    "example": "{redirect_url}?code=422&message%5Bmessage%5D=The%20provided%20credentials%20are%20incorrect.&message%5Berrors%5D%5Bemail%5D%5B0%5D=The%20provided%20credentials%20are%20incorrect"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/admin/auth/login/{provider}": {
            "get": {
                "tags": [
                    "Auth"
                ],
                "summary": "OAuth provider login endpoint",
                "operationId": "d7faf6de8b626da990c15a2ab1615605",
                "parameters": [
                    {
                        "name": "provider",
                        "in": "path",
                        "description": "OAuth provider name",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "google"
                        }
                    },
                    {
                        "name": "remember_me",
                        "in": "query",
                        "description": "Remember user",
                        "schema": {
                            "type": "integer",
                            "enum": [
                                "1",
                                "0"
                            ]
                        }
                    },
                    {
                        "name": "redirect_url",
                        "in": "query",
                        "description": "URL where the user will be redirected after authentication",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "302": {
                        "description": "Redirects to auth page"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "ValidationError": {
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "Error description"
                    },
                    "errors": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Laravel Sanctum token stateless authentication",
                "scheme": "bearer"
            },
            "sanctum_frontend": {
                "type": "apiKey",
                "in": "cookie",
                "name": "laravel_session",
                "description": "Laravel Sanctum stateful authentication"
            },
            "oauth2_google": {
                "type": "oauth2",
                "description": "Google OAuth2 stateful authentication",
                "flows": {
                    "implicit": {
                        "authorizationUrl": "/admin/auth/login/google?redirect_url=http%3A%2F%2Flocalhost&remember_me=1"
                    }
                }
            }
        }
    }
}